refactor(runtime): decouple daemon assembly from TUI
