+++ packages/cli/README.md
@@ -1,0 +1,6 @@
+<!-- SPDX-FileCopyrightText: 2026 Hari Srinivasan -->
+<!-- SPDX-License-Identifier: Apache-2.0 -->
+
+# `@axl/cli`
+
+This package owns the `axl` executable. It parses process arguments, manages local preferences and daemon connection startup, and selects the terminal client. Runtime assembly lives in `@axl/runtime`; terminal rendering lives in `@axl/tui`.
