+++ packages/cli/src/index.ts
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: 2026 Hari Srinivasan
+// SPDX-License-Identifier: Apache-2.0
+
+export * from "./settings.ts";
