+++ packages/cli/test/settings.test.ts
