+++ tsconfig.base.json
@@ -13,7 +13,9 @@
       "@axl/extension-skills": ["./packages/extensions/skills/src/index.ts"],
       "@axl/kernel": ["./packages/kernel/src/index.ts"],
       "@axl/protocol": ["./packages/protocol/src/index.ts"],
-      "@axl/sandbox": ["./packages/sandbox/src/index.ts"]
+      "@axl/runtime": ["./packages/runtime/src/index.ts"],
+      "@axl/sandbox": ["./packages/sandbox/src/index.ts"],
+      "@axl/tui": ["./packages/tui/src/index.ts"]
     },
     "lib": ["ES2023", "ESNext.Array"],
     "strict": true,
